Storefronts (#62)
* Implemented basic storefronts page

* Implemented basic css

* Create StorefrontLinkForm.js

* Create StorefrontLinkForm.css

* Update to gen additional key

Script now also generates the `.storefrontkey` to encrypt the storefront info

* Update .gitignore

* Update README.md

* Create storefronttable.go

* Create encryption.go

* update to include storefront table

* Updated home to dynamically fetch stores

* Added update function

* Update routes.go

* Added option to edit storefront

* Added delete button

The delete button also calls to eh `api/delete_storefront` to remove the storefront from the database

* Added favicons to the storefront

* change because of 'unique' requirements

because 'unique' values are enforced in the database we can no longer just save an item "over" another and must instead explicitly update the item in the database. This change in the `prodtable.go` and `prodtable_test.go` should reflect this change.

* updated to be more explicit with key

* moved clear functions to `storefrontable.go`

* Update go.mod

* Create storefronttable_test.go

* Updated Swagger Docs